Remove redundant `*ProgrammaticOptions`
