+++ catalog/files/impl/src/jmh/java/org/projectnessie/catalog/files/adls/AdlsClientResourceBench.java
@@ -41,9 +41,9 @@ import org.openjdk.jmh.annotations.Threads;
 import org.openjdk.jmh.annotations.Warmup;
 import org.openjdk.jmh.infra.Blackhole;
 import org.projectnessie.catalog.files.config.AdlsConfig;
-import org.projectnessie.catalog.files.config.AdlsProgrammaticOptions;
+import org.projectnessie.catalog.files.config.AdlsOptions;
 import org.projectnessie.catalog.files.config.ImmutableAdlsNamedFileSystemOptions;
-import org.projectnessie.catalog.files.config.ImmutableAdlsProgrammaticOptions;
+import org.projectnessie.catalog.files.config.ImmutableAdlsOptions;
 import org.projectnessie.catalog.secrets.ResolvingSecretsProvider;
 import org.projectnessie.catalog.secrets.SecretsProvider;
 import org.projectnessie.objectstoragemock.ObjectStorageMock;
@@ -82,8 +82,8 @@ public class AdlsClientResourceBench {
                           theKey, keySecret("foo").asMap())))
               .build();
 
-      AdlsProgrammaticOptions adlsOptions =
-          ImmutableAdlsProgrammaticOptions.builder()
+      AdlsOptions adlsOptions =
+          ImmutableAdlsOptions.builder()
               .defaultOptions(
                   ImmutableAdlsNamedFileSystemOptions.builder()
                       .account(URI.create("urn:nessie-secret:plain:" + theAccount))
